+++ Server/Packet.cs
@@ -92,11 +92,11 @@ public class Packet {
 
             switch((Data.TYPE)types[i]) {
                 case Data.TYPE.U8:
-                AddU8((byte)(int)Godot.GD.Convert(data[i], (int)Godot.Variant.Type.Int));
+                AddU8((byte)(int)Godot.GD.Convert(data[i], Godot.Variant.Type.Int));
                 break;
 
                 case Data.TYPE.U16:
-                AddU16((ushort)(int)Godot.GD.Convert(data[i], (int)Godot.Variant.Type.Int));
+                AddU16((ushort)(int)Godot.GD.Convert(data[i], Godot.Variant.Type.Int));
                 break;
 
                 case Data.TYPE.STRING:
